requirements.txt
Flask==3.0.0
python-dotenv==1.0.0
requests==2.31.0
flask-cors==4.0.0
gunicorn==21.2.0
gevent==23.9.1
Werkzeug==3.0.1
